Move validation under form
